Clean up

+++ src/main/java/com/everis/tddstarter/util/RomanNumeralTranslator.java
@@ -2,7 +2,7 @@ package com.everis.tddstarter.util;
 
 public class RomanNumeralTranslator {
 
-    public static String translate(Long number){
+    public static String translate(int number){
         return null;
     }
